feat: folder/filestructure scraper
