--- src/controller/characterController.ts
@@ -1,13 +1,0 @@
-import { db } from '../model/characterModel';
-import { Character } from '../model/type';
-
-async function createNewCharacter(character: Character) {
-
-}
-
-async function updateCharacterData(character: Character) {
-
-}
-
-
-export { createNewCharacter, updateCharacterData }
