+++ src/scraper/characterPage/characterScrape.js
