+++ src/scraper/frameDataPage/tableScrape.js
@@ -1,3 +1,5 @@
+import { handleImage, replaceWithAbbreviation } from './helperFunctions';
+
 function extractDataFromTable() {
   
   const rows = document.querySelectorAll('#framearea > div > table > tbody > tr');
@@ -5,38 +7,7 @@ function extractDataFromTable() {
   const results = [];
   
   let styleIndex = -1;
-  
-  const filenamesControllerImages = {
-    'arrow_3.png': '_',
-    'icon_kick.png': 'K',
-    'icon_kick_h.png': 'HK',
-    'icon_kick_l.png': 'LK',
-    'icon_kick_m.png': 'MK',
-    'icon_punch.png': 'P',
-    'icon_punch_h.png': 'HP',
-    'icon_punch_l.png': 'LP',
-    'icon_punch_m.png': 'MP',
-    'key-all.png': '*',
-    'key-d.png': '2',
-    'key-dl.png': '1',
-    'key-dr.png': '3',
-    'key-l.png': '4',
-    'key-nutral.png': '5',
-    'key-or.png': '|',
-    'key-plus.png': '+',
-    'key-r.png': '6',
-    'key-u.png': '8',
-    'key-ul.png': '7',
-    'key-ur.png': '9',
-    'modern_auto.png': 'AUTO',
-    'modern_dl.png': 'DI',
-    'modern_dp.png': 'DP',
-    'modern_h.png': 'H',
-    'modern_l.png': 'L',
-    'modern_m.png': 'M',
-    'modern_sp.png': 'SP',
-  };
-  
+
   const getText = (element) => {
     if (!element) return [];
     if (element.childNodes.length === 0) return [element.textContent];
@@ -63,27 +34,6 @@ function extractDataFromTable() {
       }
     });
     return images;
-  };
-  
-  const extractFileName = (url) => {
-    const regexSelectImgFileName = /([^\/]+)$/;
-    const match = url.match(regexSelectImgFileName);
-    return match ? match[0] : null;
-  };
-  
-  const handleImage = (element) => {
-    if (element.tagName.toLowerCase() === 'img' && element.src) {
-      const filename = extractFileName(element.src);
-      return filename ? [filename] : [];
-    }
-    return [];
-  };
-  
-  const replaceWithAbbreviation = (arrayOfFilenames) => {
-    return arrayOfFilenames.map(filename => {
-      if (!filename) return filename;
-      return filenamesControllerImages[filename] || filename;
-    });
   };
   
   rows.forEach(row => {
