+++ src/scraper/routes.ts
@@ -1,5 +1,5 @@
 import express, { Router } from 'express';
-import { scrape, scrapeOneCharacter } from './scrapeController';
+import { scrape, scrapeOneCharacter } from './controller/scrapeController';
 
 const router: Router = express.Router();
