+++ src/scraper/utils/filenameAbbreviations.js
@@ -1,0 +1,33 @@
+const filenameAbbreviations = {
+  'arrow_3.png': '_',
+  'icon_kick.png': 'K',
+  'icon_kick_h.png': 'HK',
+  'icon_kick_l.png': 'LK',
+  'icon_kick_m.png': 'MK',
+  'icon_punch.png': 'P',
+  'icon_punch_h.png': 'HP',
+  'icon_punch_l.png': 'LP',
+  'icon_punch_m.png': 'MP',
+  'key-all.png': '*',
+  'key-circle.png': 'C',
+  'key-d.png': '2',
+  'key-dl.png': '1',
+  'key-dr.png': '3',
+  'key-l.png': '4',
+  'key-nutral.png': '5',
+  'key-or.png': '|',
+  'key-plus.png': '+',
+  'key-r.png': '6',
+  'key-u.png': '8',
+  'key-ul.png': '7',
+  'key-ur.png': '9',
+  'modern_auto.png': 'AUTO',
+  'modern_dl.png': 'DI',
+  'modern_dp.png': 'DP',
+  'modern_h.png': 'H',
+  'modern_l.png': 'L',
+  'modern_m.png': 'M',
+  'modern_sp.png': 'SP',
+};
+
+export { filenameAbbreviations }
