+++ src/scraper/utils/filenameAbbreviations.json
@@ -16,7 +16,7 @@
   "icon_punch_m.png": "MP",
   "key-all.png": "*",
   "key-barrage.png": "kb",
-  "key-circle.png": "kc",
+  "key-circle.png": "C",
   "key-d.png": "2",
   "key-dl.png": "1",
   "key-dr.png": "3",
